#31 Mesh data construction
Discovered that the trailing data at the end of some meshes is actually padding for alignment. Meshes appear to be 4-byte aligned, so when serialising, this needs to be considered. LevelReader and TRMesh classes updated accordingly and unit testing is passing.
See https://www.tombraiderforums.com/archive/index.php/t-215247.html