#149 Entity limits
If the Harpoon gun is given on unarmed levels, only an additional 4 sets of harpoons will now be given (before enemy scoring multiplication), but the pistols will always be included in this instance. This avoids breaching the Entity[] limit. Additional checks are also done in ItemRandomizer to ensure this limit is not breached.