#153 Skidoo limits
The game creates black skidoos for the drivers when the level is loaded and so it needs space in Entity[] to do this. Checks are now done for this in ItemRandomizer as well as for the dragon - calls to the original CreateItem were scanned and the dragon also needs 7 slots. No other calls to this that fail seem to generate an S_ExitSystem.