#156 Flame colours
Fixes a problem with flame colours not randomizing if they are imported alongside the dragon into Tibet or Opera House (these levels don't have flames by default). This update also ensures that if the dragon and flamethrower are together in either of these levels, the sprite sequence will only be imported once.