#166 Achievement landmarks
Adds the ability to change room textures to include nods to player achievements or notable events. This is done as part of TextureRandomizer, but as this depends on importing additional textures, if the level is already at its limit (e.g. post cross-level enemy import) the import will silently fail. Source images are defined in the same way as for texture randomization. The level mapping files define the room number and rectangle indices of that room to map to the new textures.