#152 Randomized Outfits
Adds an option to OutfitRandomizer to randomly remove Lara's braid in some levels. This doesn't actually delete the model or meshes, it just remaps the textured rectangles and triangles to a transparent texture. Every level (bar the assault course) has a transparent square used by the Gunflare_H model, so we share the use of this here.