#152 Randomized Outfits
Fixes a problem when Lara's outfit is replaced by an outfit without boots. The CD player  and Lara's boots share a texture (black square), so when this was removed the CD player (specifically the headphones) in the inventory took on a replaced texture. This only affected the levels in this commit.