#174 Offshore keycard textures
The keycard textures in Offshore Rig and Diving Area will now be randomized, unless original key sprites is selected. This includes an update to StaticTextureSource to support this option as it was previously only used for dynamic textures.
Also included are updated acid, cordon and sky textures for these levels (the sky is also used in the Venice levels).