#173 Randomize in-game text
Adds core functionality to randomize text in the game. Global strings such as weapon names are defined and some grouped in gamestrings.json, so that ammo strings can match up to weapon names. Per-level definitions are also included to change level names, key names, pickup names and puzzle names. defaults.json is included purely as a reference for the indices used in the script for future changes - only gamestrings.json is used during randomization.
Care must be taken not to include any of the following characters as they are used by the game to generate accented characters and won't render as expected.
$ ) ( ~ = +