#49 Environment randomization
Added some functionality for defining environment packs, which can be used currently to add ladders and raise/lower (flat) tiles. Environment randomization takes place after item randomization (in case of item location issues with floor moves, not currently implemented) and before night randomization as floor manipulation can add vertices we would want to be darkened.
Packs can be enforced, which is the case for those currently defined. This adds a ladder and a step in GW to be able to get back to the level start after dropping through the first guardhouse.