#192 Opera/Barkhang Freeze Workaround
The cause of the freezing in Opera House and Barkhang appears to be water enemies, in particular how the game handles so many of them remaining active when floating to the top of the level. For now all water creatures are banned from both of these levels. Special thanks to Lilly Jericho and Topixtor for identifying the pattern during intensive testing.

In addition, the previous restriction of having one type of monk and one type of mercenary in Barkhang has been lifted, so any enemy except for the water creatures can now appear here.