File backup
Update in TRGE to ensure that any files copied into the backup folder have read-only flags removed. This ensures TRGE can clear away any files it doesn't need instead of throwing an exception when it tries to delete RO files. This is a workaround until the backup method itself is improved.