Refactor Zones
Rather than using ushort arrays in TR2Level, zone group objects are now used. These are built from the raw ushort array when reading the level and so adding/removing zones is just a case of manipulating the object array in TR2Level rather than changing the multidimensional arrays. TR2BoxUtilities will flatten the objects when writing the level file back.
Unit test for zones updated to match.
FlipStatus enum moved to TR base as this can be used for TR1 when the time comes, we would just need to create the matching TRZone and TRZoneGroup classes, and TRZones enum (same as TR2Zones, excluding zones 3 and 4).