#219 Initial TR3 Support
Further update to TRGE because of issues with secret counts. This is something else found to be hardcoded in the game. While TR2 allows turning secrets on/off, TR3 doesn't and you cannot change the number in any way. So TR23ScriptedLevel has now been split into TR2ScriptedLevel and TR3ScriptedLevel. Each has a read-only property to retrieve the secret count for the level. For TR2 it will always be 3 or 0. For TR3, it's linked to the level's sequence.