#219 TR3 Boxes, overlaps and zones
Updated TR3Level with the same zoning structure as TR2 and updated TR2BoxUtilities to match. TR3 is slightly different in that it allows boxes to point to the same overlap index, so this is factored into the unit testing. All tests passing for TR2 and 3.

FDUtilities updated to allow locating a room sector given coordinates. This takes into account TR3 triangles.