#219 TR3 Sequence-based modifications
Modifications will now take place based on level sequencing to cover issues with artefact pickups. If a boss level falls on Meteorite Cavern's sequence, triggers will be added under the artefacts to end the level. City is slightly different because Sophia moves around, so the pickup is moved in this instance and a path added to reach it. All changes are done using environment modifications so the JSON can be freely edited.

If Meteorite Cavern is not in its original sequence, Willard is replaced with other enemies because the only way to kill him is by picking up the stones. The stones however have to be changed to keys, otherwise picking them up ends the level immediately. Even though Willie has gone, the stones still need to be collected to open the door near the end.

Added logic to import the UPV into levels that have the exposure meter i.e. if they fall on Antarctica or RX-Tech sequence.