#231 TR3 Alternative City Ending
Because of the issue with the electrified fields and level sequencing, the alternative ending for City (used when it is in Cavern's sequence) is now used whenever the level is out of sequence. Additional crystals are added to the level to help with reaching the end trigger.
This adds an AddEntity function to the environment editor.