#232 South Pacific Spikes
The game initialises spikes in Coastal Village and Madubu to a certain height, but this is dependent on these levels being in their original sequences. If they are off-sequence, the height will be amended accordingly. For other levels that have spikes and end up in these sequences, their heights are unaffected as they don't have triggers like the South Pacific ones (barbed wire in London, Nevada, tooth spikes in India). Each "spike" level has been tested when they are in these sequences.