#237 Globe Display Options
Adds options for level sequencing to choose how to display the "next adventure" string. The options are:
- Default: the London, South Pacific, Nevada and Antarctica names will remain but there is no guarantee these match the next level.
- Area: whichever area the next level falls under will show, so while the map might highlight London, the text will say South Pacific, Nevada etc.
- Level: explicitly states the next level name.

This changes the UI layout slightly between TR2 and 3. TR2 will continue to show the number of playable levels and seed in the main window, while TR3 will show all options in the Advanced window.

The Advanced button text has been changed to "Options".