#238 Reward Room Cameras
Adds an option to show camera hints for reward room locations. Each mapping file has been updated with a camera for each room. This option will only show for TR3 in the Advanced window, TR2 will remain as before with no Advanced options in the UI and just the standard checkboxes.