#259 Tile Edge Secret Handling
If secrets are positioned within 10% of tile edges, additional triggers will be added where possible to neighbouring sectors. This allows the secret to register if Lara is on an adjoining sector rather than the original. Criteria for valid neighbours:
* Not inside a wall
* Same height as base sector
* Same RoomBelow as base sector (e.g. not in mid-air)
* Not a slope, unless the base sector is also a slope

This allows for the All Hallows secret referenced in #253 to be reinstated.