#260 TR3 Conditional Reward Rooms
Reward rooms will now be created after placing secrets, which allows for using conditions in the environment JSON to determine if a specific room is required. This is currently limited to one condition per room set, and if the condition evaluates to true for the room, it will be enforced. Otherwise, it is excluded from the available pool.
HSC mapping has been updated to use this feature.