#246 TR3 Textures
Adds the ability to replace a specific colour within tile regions with another. This is going to be needed for skyboxes for the likes of Jungle, which has "fake" sky combined into other textures (see tile 4 [128, 128]). With this approach, we can still apply dynamic mapping to the jungle, then replace the skybox, and then replace the sky colours in these specific tiles to match.