#240 TR3 Ammo for Enemy Difficulty
Adds extra ammo to unarmed levels based on the difficulty, which is calculated in a similar way to TR2. Pistols will also be added as a pickup if the level is hard or very hard (provided there is entity space and no other pistols in the level). Ammo and medi kits are added to the inventory.