#240 #241 Enemy Environment Changes
Adds basic environment changing support, mainly targeted at mods post-enemy randomization. A new group is available - ConditionalAll - each item in this set has a condition and so will only run if the condition is met (can define OnTrue and/or OnFalse). This currently fixes moves enemy 8 in Jungle if it's not suitable for the crawlspace and handles the handscanner area in HSC if there isn't a prisoner to open the trapdoor.

Environment rando will currently only apply All, ConditionalAll and NonPurist if defined. The UI options remain disabled.