#278 Flamethrower Adjustments
If flamethrowers are in Floating Islands, they won't appear in slots 34 or 35 unless enemy difficulty is set to "no restrictions". If they do end up here, environment mods will move them to the bridge so Lara can safely pull the lever. If difficulty is set to default, only the enemy directly in front of the lever can become a flamethrower.