#276 TR3 Entity Limit
Adds supports for repurposing items in levels in order to stay within the entity limit safely. An item tracker object is shared between randomizers which allows selecting reusable indices or creating new items if possible. Item rando can also provide more items to repurpose if one-limit is enabled.
Aldwych currently has 9 reusable items. Some of these are the Animating4 entities in the masonic room, so as these may be removed, environment mods will make the crawlspaces reachable. 2 secret locations have been defined here (limited to 2 because that would be the minimum removed if a level has secrets - 1 secret + 1 door).
Also updates Area51 to not include a crystal in the reward room as it had a trigger elsewhere in the level.
Item Randomizer will now also remove crystal triggers if one-limit is being used and the crystals are hidden.