#277 Artefact Menu Models
Ensures that artefacts carried over from previous levels will be visible in the inventory. A side effect of this is that texture randomizer will change the artefact colours if they match one that is being used for secrets in the level, but this would only affect "previous" artefacts - standard artefact pickups in the level will remain default.