#245 TR3 Start Positions
Start positions for TR3 implemented. This differs slightly from TR2 as we move any triggers that are below Lara to her new position. Some levels only have a handful of locations because of the trigger types nearby. This also updates the EM move trigger function with some defences such as checking the base and target sectors aren't the same.