#286 TR3 Enemy and Heavy Trigger Handling
All instances where an enemy shares a sector with an AIAmbush entity and corresponding nearby heavy triggers have been checked and workarounds put in place for when the enemy isn't suitable. The main issue is with monkeys as they want to drop a medi-pack when they reach the destination ambush point, but they won't always have one, and this can cause a crash or strange behaviour (for example in room 31 in Aldwych, when the monkey hits the heavy trigger, the button and door disappear and the monkey voids). Other instances are where enemies can't reach the destination point (e.g. crawlers if they have to go up a step) so again, workarounds are provided.
Most of the handling is in the environment JSON, but some additional checks are performed in the enemy randomizer itself. In addition, improvements to ensuring that a "safe" enemy is available is now provided, so in the rare chance that a randomization pool contains all restricted enemies, we can guarantee also a non-restricted one to ensure the placement rules aren't breached.
Willard has also been banned from All Hallows unless difficulty is set to no restrictions.