#289 Shiva Dependencies
Shiva depends on the monkey and monkey mesh swap models being available, but this only seems to be the case when two are triggered at the same time or when one is triggered while another is turning back into stone. TR3 model imports will now sort the models by ID to replicate order in original levels (not implemented for TR2 to avoid introducing any issues there).