#288 TR3 Key Item Triggers
Trigger handling for key items has been moved from TR3ItemRandomizer to default conditional environment checks. So if for example room 147 in Temple Ruins does not contain the Ganesha key in its usual position, the original pickup trigger for the door is converted into a pad to avoid a softlock.