#292 Advanced Environment Conditions
Adds the ability to build more complex conditions in environment JSON using negation plus and/or/xor combinations. This has been applied for TR2 as an example to show that an enemy can be created based on a room's water status and the model being available.
Also fixes a potential null reference exception in the environment randomizers if only defining either or OnTrue/OnFalse in conditional sets.