#295 TR3 Mirrored Secret Filtering
Adds a LevelState property to Location to allow secrets to be tied to mirrored or non-mirrored levels. The default is Any meaning they are universal. If mirroring is selected, the list of levels will be allocated prior to the secrets being placed so that the secret randomizer knows which locations it can use (the mirroring itself still takes place later, so this also means that mirrored-only secrets must be defined in normal world space).

Added a randomizer.json trview settings file to the shared resources for convenience. This won't be included in the build.