#299 HSH Texture Amendments
The embedded HSH file has been updated to free-up texture space, meaning that enemies that were previously restricted because of texture size can now appear in this level. This also adds a new method of defining special case enemies where we define known combinations that work against each difficulty, so that if the enemy is picked, one of these combinations will be picked.