#310 Multiple Dragons
Improvement to handling of Marco in Offshore Rig. If environment rando is not enabled, the plane area will never be drained, so this case is now handled (i.e. Marco's appearance will be undone).
This adjusts environment rando by applying ConditionalAll after any other mods so that conditions can be based on aspects that may have changed here.