#349 TR1 Model Transport
Fixes a problem if importing the centaur statue into a level that already has the centaur (a dependency). Previous dependency checks assume the parent model would already be present, so the bug was that the dependency would be removed.
Also fixes explosion sprite in Vilcabamba (it already exists there but is empty).