#361 TR1 Audio Randomization
Audio randomizer implemented, which includes a change to the shared audio class for all 3 randomizers. This resolves an issue when setting trigger tracks if a group of triggers spans more than one room (good example is the T-Rex trigger in Lost Valley). The previous implementation was grouping tracks by room, so it was possible here to end up with different tracks. Trigger groups are now detected and handled.