#354 TR1 Item Randomization
Adds an option to allocate additional pickups. This is currently targeted at early levels (up to and including Midas) as the number of pickups is low and so can be challenging with enemy randomization.
The option is not added as a separate mode alongside Default/OneLimit so that it can be used with either of these options.