#349 TR1 Model Transport
Correction for sample import - this was using the TR2/3 check of ensuring duplicate indices were not imported but this doesn't apply to TR1 as all the WAV data is in the level file. So while this largely worked, for some cases the map/details would be imported but the WAV data wouldn't be.