#361 TR1 Audio Randomization
Avoids a potential issue is a sound ID is not updated in the sound map, but points to a sound details object with more samples than the one we are trying to import. This results in importing more WAV data, but resorting the sounds will tidy up the data regardless once complete.