#353 TR1 Secret Randomization
Allows for 3D models to be used even if randomizing secrets (provided "use recommended settings" is disabled). The issue was with the scion as it becomes embedded in the ground, so a workaround is put in place to raise the mesh (so the item's Y position remains consistent). The sprite will also now appear sitting on the ground.