#360 TR1 Environment Randomization
Added the ability to generate room vertex lighting dynamically. This is kept separate from creating rooms as we may wish to add or amend faces before lighting is set. It currently locates the nearest room light to each vertex and applies a lighting value based on the distance and the light's fade.