#356 TR1 Texture Randomization
If a mesh texture is used on a wall, and it's not 64x64, packing may end up allocating the mesh texture (which doesn't use anti-aliasing in wireframe mode) and so this results in all walls not using anti-alias. Proper fix is to amend texture packing to allow for this scenario, but for now the room size frame will be normalised to 64x64.