Update Diving Area environment JSON
Replaced the old return path in Diving Area (lots of ladders) with some rooms to make it easier for the player.
Added another ladder to rooms 3/4.
Fixed the light generating function for TR2.
Changed the room references if a dragon room has been imported to be end-relative rather than fixed.