Update TR1 environment changes
Update to door mirroring to avoid treating doors that are in front of each other as side-by-side double doors. This only affected the ones at the end of Vilcabamba, which wasn't too big an issue before, but it would be possible with the coming key item changes for the gold idol to appear between the doors, so it would lead to a softlock.

General environment updates:
City of Khamoon - makes sure a door is open after leaving a challenge room, as key items may end up here.
Obelisk - a mod may add a key to exit the level. We now leave this in a default spot rather than picking one from the mods, and key item rando will detect and move it.
Mines - return paths added in the TNT area. And similar to Obelisk, we make an added key item static for key item rando to detect.