Fix TR2 model adjustments
TR2 replaces P2 items with P3, to allow for the dragon to be placed in each level (the dagger is hardcoded as a P2 item). This was always done after key item rando to allow for zoning of the original items. I forgot to move this logic in #568. Zoning of these items must now be done based on them being P3.
Removed the puzzle reassignment checks from TR3 as it's not applicable there.
Also updated the enemy adjustment logic (this clears out unused enemies in heavy-item levels) so that the items can be re-used if needed.