Create unified zoning logic
Key items and secrets will now be placed using identical logic in all games. This uses routes (currently empty) to mark the minimum and maximum rooms key items can be placed in to prevent softlocks. For secrets, the previous sphere-testing method is replaced with a distance approximator, again using the provided level route. TR2 secrets will continue to be in Stone/Jade/Gold order, based on the dev-intended route of a level.