Route all levels
Routing created for every level in TR1-3. This makes use of a trview plugin to allow visualisation of the created zones, which are generated by marking on waypoints low and high rooms along the route for each key item.
Updated invalid item locations too as more came to light during this exercise.