Refactor FDControl
TRUB has packed floor data, so our current method was failing as it always expected each sector to point to a unique entry. I've streamlined the class overall, removing lots of duplication, and updated the tests to run against every level. We can't test byte-for-byte in TRUB, so we ensure that each sector continues to point to a valid entry. There was also one snag with Antarctica in TR3 which has a single unused triangulation entry, so there is a workaround in the test to get by this. GIGO.