Add enemy rando support
A tweak to enemy randomization to ensure we don't write back floor data too early while we're still scanning over it.
Texture dependencies calculated for model import.
Some room restrictions defined for Adam in default mode.
Locations defined for eggs.
Adjusted the enemy key drop option to allow keys to be shifted if the player wants to play pacifist. For TR1, Pierre's items in Tihocan are adjusted to add them to the level for trview lookup.