Additional puzzle room
Added functionality to create custom textures, and to lock those textures to avoid wireframing replacing them. This is used in a puzzle room in The Hive where texture hints are needed. This also uses a mechanism to lock audio tracks, as again this room uses audio hints so we want to guarantee specific tracks.

More boulders added to Stronghold.