Remove unused uint
All levels store a uint before the room count which is always 0 and is ignored in all games during read, It is believed to be a level number field. We store this for 1-3 in case it becomes usable, TR4/5 is a little more awkward because of where it's kept in the compressed part of the file, so for now we just keep it in place.