Fix alias priorities
This only affected TRR because of the way we merge the PDP files, but ultimately this performs a better check on alias priorities before import i.e. if the level already has an alias of a type and it's the "chosen one", remove the other's blob before proceeding.
TR3 aliases for Lara misc are ignored because none contain critical commands like ending the level.