Fix key item issues
Consider enemies 71/72 in Wreck - if they are allocated either of the first two keys, it's a softlock. While the room is valid en-route, the enemies' trigger is beyond the keyhole. For shuffled mode, it's too complicated to re-allocate, so we just disable key drops in that scenario. For normal mode, the triggers are taken into consideration.