Fix bird monster commands
Imported bird monsters for TR2X will not have the end level animation command (to avoid potential conflict in TR2R) so this ensures the command is either added or removed based on the chosen setting. Additionally, docile bird monsters (monks) will not inherit the command if it is there by default, such as in Ice Palace or Vegas.